Add cluster with iterative variable selection

This function adds cluster assignment to each row of a data.frame with iterative variable selection during clustering.

add_cluster_assignment(
  df,
  v_cluster = NULL,
  k = NULL,
  maxit = 10,
  return_df_cluster_instead = FALSE,
  new_var_name = "Cluster"
)

Arguments

df

data.frame

v_cluster

(character) vector of variable names to use in clustering (if NULL, use all)

k

(numeric) number of clusters (if NULL, determined optimally; see get_cluster())

maxit

(numeric) maximum number of iterations to select significant outcomes of clusters

return_df_cluster_instead

(logical) whether to return df with only final clustering variables and cluster assignment (default FALSE)

new_var_name

(character) new variable name

Value

df with new cluster variable added

Details

An initial set of clusters is determined using get_cluster() (Gower distance matrix with clustering around medoids). Then, significant differences between clusters are determined using get_sig_differences_between_groups() (ANOVA for numeric variables, Chi-square for categorical variables), and clustering is performed again using the reduced set of significant variables. This variable-reduction method is repeated until all remaining variables are significant, or until maxit is reached. If maxit is reached before convergence, a warning is thrown.

When maxit is set to 0, a warning is thrown indicating that no variable selection is performed, and this is equivalent to using get_cluster() directly.

Examples

if (FALSE) {
df <- df |> add_cluster_assignment()

df |>
   add_cluster_assignment(return_df_cluster_instead=TRUE) |>
   plot_density_by_groups(group="Cluster")
}